fix(security): Check url

+++ komitet_git_bezopasnosti/github.py
@@ -10,6 +10,14 @@ from .config import STATUS_CONTEXT
 log = logging.getLogger(__name__)
 
 
+def check_url(url, start="https://api.github.com/"):
+    if url.startswith(start):
+        return True
+    log.error({"error": "UNTRUSTED_API_URL",
+               "bad_url": url})
+    return False
+
+
 def protect(string):
     return string.replace("[", "\[")
 
@@ -19,24 +27,32 @@ def quote(string):
 
 
 def get_commits(url):
-    commits = requests.get(url, params={"access_token": GH_TOKEN})
-    return commits.json()
+    if check_url(url):
+        commits = requests.get(url, params={"access_token": GH_TOKEN})
+        return commits.json()
+    else:
+        return None
 
 
 def get_comments(url):
-    comments = requests.get(url, params={"access_token": GH_TOKEN})
     result = []
-    for comment in comments.json():
-        if comment["body"].startswith(HIDDEN):
-            result.append(comment)
+    if check_url(url):
+        comments = requests.get(url, params={"access_token": GH_TOKEN})
+        for comment in comments.json():
+            if comment["body"].startswith(HIDDEN):
+                result.append(comment)
     return result
 
 
 def delete_comment(comment):
-    requests.delete(comment["url"], params={"access_token": GH_TOKEN})
+    url = comment["url"]
+    if check_url(url):
+        requests.delete(comment["url"], params={"access_token": GH_TOKEN})
 
 
 def upsert_comment(url, messages):
+    if not check_url(url):
+        return
     comments = get_comments(url)
     message = HIDDEN + "\n\n".join(messages)
 
@@ -56,20 +72,23 @@ def upsert_comment(url, messages):
 
 
 def update_comment(comments, body):
-    requests.patch(comments["url"],
-                   params={"access_token": GH_TOKEN},
-                   json={"body": body})
+    if check_url(comments["url"]):
+        requests.patch(comments["url"],
+                       params={"access_token": GH_TOKEN},
+                       json={"body": body})
 
 
 def create_comment(url, body):
-    requests.post(url,
-                  params={"access_token": GH_TOKEN},
-                  json={"body": body})
+    if check_url(url):
+        requests.post(url,
+                      params={"access_token": GH_TOKEN},
+                      json={"body": body})
 
 
 def update_status(url, state, message):
-    requests.post(url,
-                  params={"access_token": GH_TOKEN},
-                  json={"state": state,
-                        "context": STATUS_CONTEXT,
-                        "description": message})
+    if check_url(url):
+        requests.post(url,
+                      params={"access_token": GH_TOKEN},
+                      json={"state": state,
+                            "context": STATUS_CONTEXT,
+                            "description": message})
